requirements.txt
amqp==2.6.0
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.6
certifi==2020.6.20
chardet==3.0.4
Django==3.0.8
djangorestframework==3.11.0
future==0.18.2
idna==2.10
importlib-metadata==1.7.0
kombu==4.6.11
pkg-resources==0.0.0
psycopg2-binary==2.8.5
pyshorteners==1.0.1
pytz==2020.1
requests==2.24.0
short-url==1.2.2
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.9
vine==1.3.0
zipp==3.1.0
